Fixed Crash on Safari
On Safari, for some strange reason, the variable layers cannot be found. This is solved by moving the layers variable to the top of the class. In addition, on some versions of Safari, gamepads are not supported, and the navigator.getGamepads() function is undefined. This is solved by verifying that the function exists before calling it preventing a crash.