Navigation bar fades and reappears
"TopButtonContainer" is visible for 10 seconds before it fades and shrinks, allowing the user to have plenty of time to assess the buttons before they disappear. If the user so wishes to access one of the buttons, a mouseover at the top of the display will fade the bar back in. It will fade again after another 10 secs. This adds a general polish to the game and makes the top bar less obtrusive during gameplay. Nothing too complex ;)